hostirique facturation

+++ database/migrations/2017_11_15_123730_create_facture_table.php
@@ -20,7 +20,7 @@ class CreateFactureTable extends Migration
 			$table->string('numero');
 			$table->mediumInteger('situation');
 			$table->string('type');
-			$table->tinyInteger('client_id');
+			$table->integer('client_id');
 			$table->string('nom_client');
 			$table->string('adresse');
 			$table->string('adresse_comp')->nullable();
